# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/events": {
            "get": {
                "description": "Lista todos los eventos almacenados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.GenericResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un evento existente con los datos proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualiza un evento existente",
                "parameters": [
                    {
                        "description": "Datos del evento a actualizar, el valor de Status debe ser (Sin Revisar o Revisar)",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento actualizado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo evento con los datos proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Crea un nuevo evento",
                "parameters": [
                    {
                        "description": "Datos del evento a crear",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Event"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Evento creado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            }
        },
        "/events/get-management-required/{value}": {
            "get": {
                "description": "Obtiene una lista de eventos filtrados por requerimiento de gestión (true o false).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtiene eventos filtrados por requerimiento de gestión",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Valor de requerimiento de gestión (true o false)",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eventos filtrados por requerimiento de gestión",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.GenericResponse"
                            }
                        }
                    }
                }
            }
        },
        "/events/update-status": {
            "put": {
                "description": "Actualiza el estado de un evento existente con los datos proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualiza el estado de un evento",
                "parameters": [
                    {
                        "description": "Datos del estado a actualizar,  el valor de Status debe ser (Sin Revisar o Revisar)",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Estado del evento actualizado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "delete": {
                "description": "Elimina un evento existente según su ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Elimina un evento por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del evento a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento eliminado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            }
        },
        "/types-event": {
            "get": {
                "description": "Obtiene una lista de tipos de eventos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos Evento"
                ],
                "summary": "Obtiene tipos de eventos",
                "responses": {
                    "200": {
                        "description": "Tipos de eventos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.GenericResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un tipo de evento existente con los datos proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos Evento"
                ],
                "summary": "Actualiza un tipo de evento existente",
                "parameters": [
                    {
                        "description": "Datos del tipo de evento a actualizar",
                        "name": "typeEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.TypeEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de evento actualizado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo tipo de evento con los datos proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos Evento"
                ],
                "summary": "Crea un nuevo tipo de evento",
                "parameters": [
                    {
                        "description": "Datos del tipo de evento a crear",
                        "name": "typeEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.TypeEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tipo de evento creado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            }
        },
        "/types-event/{id}": {
            "delete": {
                "description": "Elimina un tipo de evento existente según su ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos Evento"
                ],
                "summary": "Elimina un tipo de evento por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tipo de evento a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de evento eliminado con éxito",
                        "schema": {
                            "$ref": "#/definitions/response.GenericResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Event": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type_event_id": {
                    "type": "integer"
                }
            }
        },
        "entity.TypeEvent": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "management_required": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "request.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.GenericResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}